prefix the overflow

+++ 2.0/src/xtn_inner.cu
@@ -100,7 +100,7 @@ int postprocessing(Int3* seq, Int2* input, int distance,
 
 	// cal levenshtein
 	int uniqueLen = transfer_last_element(buffer, 1); gpuerr();
-	int byteRequirement = sizeof(char) * uniqueLen;
+	size_t byteRequirement = sizeof(char) * uniqueLen;
 	cudaMalloc(&flags, byteRequirement); gpuerr();
 	cudaMalloc(&uniqueDistances, byteRequirement); gpuerr();
 	cudaMalloc(&distanceOutput, byteRequirement); gpuerr();
